Image/requirements.txt
fastapi
uvicorn
mysql-connector-python
minio
pydantic
